# internal/api/dronefleet.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for DroneStatus.
const (
	DroneStatusActive   DroneStatus = "active"
	DroneStatusCharging DroneStatus = "charging"
	DroneStatusFault    DroneStatus = "fault"
	DroneStatusIdle     DroneStatus = "idle"
)

// Valid indicates whether the value is a known member of the DroneStatus enum.
func (e DroneStatus) Valid() bool {
	switch e {
	case DroneStatusActive:
		return true
	case DroneStatusCharging:
		return true
	case DroneStatusFault:
		return true
	case DroneStatusIdle:
		return true
	default:
		return false
	}
}

// Defines values for MissionStatus.
const (
	MissionStatusAborted   MissionStatus = "aborted"
	MissionStatusActive    MissionStatus = "active"
	MissionStatusCompleted MissionStatus = "completed"
	MissionStatusPlanned   MissionStatus = "planned"
)

// Valid indicates whether the value is a known member of the MissionStatus enum.
func (e MissionStatus) Valid() bool {
	switch e {
	case MissionStatusAborted:
		return true
	case MissionStatusActive:
		return true
	case MissionStatusCompleted:
		return true
	case MissionStatusPlanned:
		return true
	default:
		return false
	}
}

// ConfigResponse defines model for ConfigResponse.
type ConfigResponse struct {
	AppName            string `json:"appName"`
	DatabaseConfigured bool   `json:"databaseConfigured"`
	Environment        string `json:"environment"`
	FeatureCacheWarm   bool   `json:"featureCacheWarm"`
	LogLevel           string `json:"logLevel"`
}

// Drone defines model for Drone.
type Drone struct {
	Altitude  float64     `json:"altitude"`
	Battery   int         `json:"battery"`
	Id        string      `json:"id"`
	Lat       float64     `json:"lat"`
	Lng       float64     `json:"lng"`
	MissionId *string     `json:"missionId,omitempty"`
	Name      string      `json:"name"`
	Status    DroneStatus `json:"status"`
}

// DroneMessageResponse defines model for DroneMessageResponse.
type DroneMessageResponse struct {
	Drone   Drone  `json:"drone"`
	Message string `json:"message"`
}

// DroneStatus defines model for DroneStatus.
type DroneStatus string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error string `json:"error"`
}

// FleetSummary defines model for FleetSummary.
type FleetSummary struct {
	Active int     `json:"active"`
	Drones []Drone `json:"drones"`
	Faults int     `json:"faults"`
	Total  int     `json:"total"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Service string `json:"service"`
	Status  string `json:"status"`
	Version string `json:"version"`
}

// Mission defines model for Mission.
type Mission struct {
	CompletedAt *time.Time    `json:"completedAt,omitempty"`
	DroneId     string        `json:"droneId"`
	Id          string        `json:"id"`
	StartedAt   *time.Time    `json:"startedAt,omitempty"`
	Status      MissionStatus `json:"status"`
	Waypoints   []Waypoint    `json:"waypoints"`
}

// MissionList defines model for MissionList.
type MissionList struct {
	Missions []Mission `json:"missions"`
	Total    int       `json:"total"`
}

// MissionStatus defines model for MissionStatus.
type MissionStatus string

// ReadyResponse defines model for ReadyResponse.
type ReadyResponse struct {
	Database     string `json:"database"`
	DronesLoaded *int   `json:"dronesLoaded,omitempty"`
	Status       string `json:"status"`
}

// StatusResponse defines model for StatusResponse.
type StatusResponse struct {
	Status string `json:"status"`
}

// TaskResponse defines model for TaskResponse.
type TaskResponse struct {
	ExecutedAt time.Time `json:"executedAt"`
	Status     string    `json:"status"`
	Task       string    `json:"task"`
}

// TelemetryUpdate defines model for TelemetryUpdate.
type TelemetryUpdate struct {
	Altitude float64 `json:"altitude"`
	Battery  int     `json:"battery"`
	Lat      float64 `json:"lat"`
	Lng      float64 `json:"lng"`
}

// VersionResponse defines model for VersionResponse.
type VersionResponse struct {
	AppName     string `json:"appName"`
	BuildTime   string `json:"buildTime"`
	Environment string `json:"environment"`
	GitSha      string `json:"gitSha"`
	Version     string `json:"version"`
}

// Waypoint defines model for Waypoint.
type Waypoint struct {
	Alt float64 `json:"alt"`
	Lat float64 `json:"lat"`
	Lng float64 `json:"lng"`
}

// DroneId defines model for DroneId.
type DroneId = string

// MissionId defines model for MissionId.
type MissionId = string

// UpdateDroneTelemetryJSONRequestBody defines body for UpdateDroneTelemetry for application/json ContentType.
type UpdateDroneTelemetryJSONRequestBody = TelemetryUpdate

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /api/v1/fleet)
	ListFleet(w http.ResponseWriter, r *http.Request)

	// (GET /api/v1/fleet/{droneId})
	GetDrone(w http.ResponseWriter, r *http.Request, droneId DroneId)

	// (POST /api/v1/fleet/{droneId}/recover)
	RecoverDrone(w http.ResponseWriter, r *http.Request, droneId DroneId)

	// (POST /api/v1/fleet/{droneId}/simulate-fault)
	SimulateDroneFault(w http.ResponseWriter, r *http.Request, droneId DroneId)

	// (POST /api/v1/fleet/{droneId}/telemetry)
	UpdateDroneTelemetry(w http.ResponseWriter, r *http.Request, droneId DroneId)

	// (GET /api/v1/missions)
	ListMissions(w http.ResponseWriter, r *http.Request)

	// (GET /api/v1/missions/{missionId})
	GetMission(w http.ResponseWriter, r *http.Request, missionId MissionId)

	// (GET /config)
	GetConfig(w http.ResponseWriter, r *http.Request)

	// (GET /health)
	GetHealth(w http.ResponseWriter, r *http.Request)

	// (GET /ready)
	GetReady(w http.ResponseWriter, r *http.Request)

	// (POST /tasks/cache-warm)
	WarmCache(w http.ResponseWriter, r *http.Request)

	// (GET /version)
	GetVersion(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (GET /api/v1/fleet)
func (_ Unimplemented) ListFleet(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/v1/fleet/{droneId})
func (_ Unimplemented) GetDrone(w http.ResponseWriter, r *http.Request, droneId DroneId) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /api/v1/fleet/{droneId}/recover)
func (_ Unimplemented) RecoverDrone(w http.ResponseWriter, r *http.Request, droneId DroneId) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /api/v1/fleet/{droneId}/simulate-fault)
func (_ Unimplemented) SimulateDroneFault(w http.ResponseWriter, r *http.Request, droneId DroneId) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /api/v1/fleet/{droneId}/telemetry)
func (_ Unimplemented) UpdateDroneTelemetry(w http.ResponseWriter, r *http.Request, droneId DroneId) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/v1/missions)
func (_ Unimplemented) ListMissions(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/v1/missions/{missionId})
func (_ Unimplemented) GetMission(w http.ResponseWriter, r *http.Request, missionId MissionId) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /config)
func (_ Unimplemented) GetConfig(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /health)
func (_ Unimplemented) GetHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /ready)
func (_ Unimplemented) GetReady(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /tasks/cache-warm)
func (_ Unimplemented) WarmCache(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /version)
func (_ Unimplemented) GetVersion(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListFleet operation middleware
func (siw *ServerInterfaceWrapper) ListFleet(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListFleet(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetDrone operation middleware
func (siw *ServerInterfaceWrapper) GetDrone(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "droneId" -------------
	var droneId DroneId

	err = runtime.BindStyledParameterWithOptions("simple", "droneId", chi.URLParam(r, "droneId"), &droneId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "droneId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetDrone(w, r, droneId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RecoverDrone operation middleware
func (siw *ServerInterfaceWrapper) RecoverDrone(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "droneId" -------------
	var droneId DroneId

	err = runtime.BindStyledParameterWithOptions("simple", "droneId", chi.URLParam(r, "droneId"), &droneId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "droneId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RecoverDrone(w, r, droneId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SimulateDroneFault operation middleware
func (siw *ServerInterfaceWrapper) SimulateDroneFault(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "droneId" -------------
	var droneId DroneId

	err = runtime.BindStyledParameterWithOptions("simple", "droneId", chi.URLParam(r, "droneId"), &droneId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "droneId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SimulateDroneFault(w, r, droneId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateDroneTelemetry operation middleware
func (siw *ServerInterfaceWrapper) UpdateDroneTelemetry(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "droneId" -------------
	var droneId DroneId

	err = runtime.BindStyledParameterWithOptions("simple", "droneId", chi.URLParam(r, "droneId"), &droneId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "droneId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateDroneTelemetry(w, r, droneId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListMissions operation middleware
func (siw *ServerInterfaceWrapper) ListMissions(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListMissions(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetMission operation middleware
func (siw *ServerInterfaceWrapper) GetMission(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "missionId" -------------
	var missionId MissionId

	err = runtime.BindStyledParameterWithOptions("simple", "missionId", chi.URLParam(r, "missionId"), &missionId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "missionId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetMission(w, r, missionId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetConfig operation middleware
func (siw *ServerInterfaceWrapper) GetConfig(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetConfig(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetReady operation middleware
func (siw *ServerInterfaceWrapper) GetReady(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetReady(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// WarmCache operation middleware
func (siw *ServerInterfaceWrapper) WarmCache(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.WarmCache(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetVersion operation middleware
func (siw *ServerInterfaceWrapper) GetVersion(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetVersion(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/fleet", wrapper.ListFleet)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/fleet/{droneId}", wrapper.GetDrone)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/fleet/{droneId}/recover", wrapper.RecoverDrone)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/fleet/{droneId}/simulate-fault", wrapper.SimulateDroneFault)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/fleet/{droneId}/telemetry", wrapper.UpdateDroneTelemetry)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/missions", wrapper.ListMissions)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/missions/{missionId}", wrapper.GetMission)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/config", wrapper.GetConfig)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.GetHealth)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/ready", wrapper.GetReady)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/tasks/cache-warm", wrapper.WarmCache)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/version", wrapper.GetVersion)
	})

	return r
}