# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"database/sql/driver"
	"fmt"

	"github.com/jackc/pgx/v5/pgtype"
)

type DroneStatus string

const (
	DroneStatusActive   DroneStatus = "active"
	DroneStatusIdle     DroneStatus = "idle"
	DroneStatusCharging DroneStatus = "charging"
	DroneStatusFault    DroneStatus = "fault"
)

func (e *DroneStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = DroneStatus(s)
	case string:
		*e = DroneStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for DroneStatus: %T", src)
	}
	return nil
}

type NullDroneStatus struct {
	DroneStatus DroneStatus `json:"drone_status"`
	Valid       bool        `json:"valid"` // Valid is true if DroneStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullDroneStatus) Scan(value interface{}) error {
	if value == nil {
		ns.DroneStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.DroneStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullDroneStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.DroneStatus), nil
}

type MissionStatus string

const (
	MissionStatusPlanned   MissionStatus = "planned"
	MissionStatusActive    MissionStatus = "active"
	MissionStatusCompleted MissionStatus = "completed"
	MissionStatusAborted   MissionStatus = "aborted"
)

func (e *MissionStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = MissionStatus(s)
	case string:
		*e = MissionStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for MissionStatus: %T", src)
	}
	return nil
}

type NullMissionStatus struct {
	MissionStatus MissionStatus `json:"mission_status"`
	Valid         bool          `json:"valid"` // Valid is true if MissionStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullMissionStatus) Scan(value interface{}) error {
	if value == nil {
		ns.MissionStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.MissionStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullMissionStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.MissionStatus), nil
}

type Drone struct {
	ID        string             `json:"id"`
	Name      string             `json:"name"`
	Status    DroneStatus        `json:"status"`
	Battery   int32              `json:"battery"`
	Lat       float64            `json:"lat"`
	Lng       float64            `json:"lng"`
	Altitude  float64            `json:"altitude"`
	MissionID pgtype.Text        `json:"mission_id"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type Mission struct {
	ID          string             `json:"id"`
	DroneID     string             `json:"drone_id"`
	Status      MissionStatus      `json:"status"`
	StartedAt   pgtype.Timestamptz `json:"started_at"`
	CompletedAt pgtype.Timestamptz `json:"completed_at"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type MissionWaypoint struct {
	MissionID string  `json:"mission_id"`
	Sequence  int32   `json:"sequence"`
	Lat       float64 `json:"lat"`
	Lng       float64 `json:"lng"`
	Alt       float64 `json:"alt"`
}

type TelemetryEvent struct {
	ID         int64              `json:"id"`
	DroneID    string             `json:"drone_id"`
	Lat        float64            `json:"lat"`
	Lng        float64            `json:"lng"`
	Altitude   float64            `json:"altitude"`
	Battery    int32              `json:"battery"`
	RecordedAt pgtype.Timestamptz `json:"recorded_at"`
}